Week 13 task completed

+++ src/main/java/com/example/gamefusion/Repository/BrandLogoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface BrandLogoRepository extends JpaRepository<BrandLogo,Integer> {
+public interface BrandLogoRepository extends JpaRepository<BrandLogo,Long> {
     Optional<BrandLogo> findByName(String fileName);
 }
